src/test/javascript/spec/app/entities/kontrachent/kontrachent.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { KontrachentService } from 'app/entities/kontrachent/kontrachent.service';
import { IKontrachent, Kontrachent } from 'app/shared/model/kontrachent.model';

describe('Service Tests', () => {
  describe('Kontrachent Service', () => {
    let injector: TestBed;
    let service: KontrachentService;
    let httpMock: HttpTestingController;
    let elemDefault: IKontrachent;
    let expectedResult: IKontrachent | IKontrachent[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(KontrachentService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new Kontrachent(0, 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', 0);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a Kontrachent', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new Kontrachent()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a Kontrachent', () => {
        const returnedFromService = Object.assign(
          {
            nazwaKontrachenta: 'BBBBBB',
            emailKontrachenta: 'BBBBBB',
            numerKontrachenta: 'BBBBBB',
            terminKontrachenta: 1
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of Kontrachent', () => {
        const returnedFromService = Object.assign(
          {
            nazwaKontrachenta: 'BBBBBB',
            emailKontrachenta: 'BBBBBB',
            numerKontrachenta: 'BBBBBB',
            terminKontrachenta: 1
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a Kontrachent', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
